merge ad 2020-04-02

+++ redisson/src/main/java/org/redisson/api/RBloomFilter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Nikita Koksharov
+ * Copyright (c) 2013-2020 Nikita Koksharov
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,7 @@
 package org.redisson.api;
 
 /**
- * Bloom filter based on 64-bit hash derived from 128-bit hash (xxHash + FarmHash).
- *
- * Code parts from Guava BloomFilter
+ * Distributed implementation of Bloom filter based on Highway 128-bit hash.
  *
  * @author Nikita Koksharov
  *
@@ -26,8 +24,22 @@ package org.redisson.api;
  */
 public interface RBloomFilter<T> extends RExpirable {
 
+    /**
+     * Adds element
+     * 
+     * @param object - element to add
+     * @return <code>true</code> if element has been added successfully
+     *         <code>false</code> if element is already present
+     */
     boolean add(T object);
 
+    /**
+     * Check for element present
+     * 
+     * @param object - element
+     * @return <code>true</code> if element is present
+     *         <code>false</code> if element is not present
+     */
     boolean contains(T object);
 
     /**
@@ -35,19 +47,42 @@ public interface RBloomFilter<T> extends RExpirable {
      * calculated from <code>expectedInsertions</code> and <code>falseProbability</code>
      * Stores config to Redis server.
      *
-     * @param expectedInsertions - expected amount of insertions
+     * @param expectedInsertions - expected amount of insertions per element
      * @param falseProbability - expected false probability
      * @return <code>true</code> if Bloom filter initialized
      *         <code>false</code> if Bloom filter already has been initialized
      */
     boolean tryInit(long expectedInsertions, double falseProbability);
 
+    /**
+     * Returns expected amount of insertions per element.
+     * Calculated during bloom filter initialization. 
+     * 
+     * @return expected amount of insertions per element
+     */
     long getExpectedInsertions();
 
+    /**
+     * Returns false probability of element presence. 
+     * Calculated during bloom filter initialization.
+     * 
+     * @return false probability of element presence
+     */
     double getFalseProbability();
 
+    /**
+     * Returns number of bits in Redis memory required by this instance
+     * 
+     * @return number of bits
+     */
     long getSize();
 
+    /**
+     * Returns hash iterations amount used per element. 
+     * Calculated during bloom filter initialization. 
+     * 
+     * @return hash iterations amount
+     */
     int getHashIterations();
 
     /**
@@ -55,6 +90,6 @@ public interface RBloomFilter<T> extends RExpirable {
      *
      * @return probabilistic number of elements
      */
-    int count();
+    long count();
 
 }
